soapy: suffix module targets to avoid name collisions
The module targets could overlap with the name of the module's underlying support library. This shouldnt be an issue, but LoadLibrary() on windows is bombing out with error 193 until the file with the clashing name is changed. For a quick fix, the target/library names have been suffixed with the word Module.