soapy: Support /usr/local module installs when root is /usr
This change supports searching /usr/local when /usr is the install root.
In addition, the module util will not install /usr unless specified.
This is designed to support soapy sdr in a packaged unix environment,
without violating FHS so that user-compiled modules are installed
to a non /usr install prefix, unless explicitly configured.