debian: fix CMAKE_INSTALL_RPATH detection
The logic should not have set the install RPATH, but it could not handle the multi-arch directory.
This change adds a CMAKE_AUTOSET_INSTALL_RPATH option which defaults to TRUE, but the rules set to FALSE.
Now the build RPATH can be enabled and the unit tests can be run with the in-tree library.