fix size_t python issue on arm 32-bit
Swig cant handle the case where vector<size_t> coexists
with vector<X> where size_t is a typedef of that type X.
In this case, we use vector of size_t and unsigned
which are the same type on arm 32-bit gcc compilers.
The solution is to detect the condition when size_t is unsigned
and to ifdef-ify the swig definitions when this happens.