set soapysdr abi version to 0.7 for recent additions
Although the core device ABI has not changed on the development branch,
modules may use the new converter API or verion registration API.
This creates a compatibility mismatch for modules built on the
development branch vs the previous release. Therefore, modules built
against this branch will be installed in the new modules directory for 0.7.