python3 - change OUTPUT_NAME in newer versions of cmake
From UseSWIG:

Changed in version 3.15: Alternate library name (set with the OUTPUT_NAME property, for example)
will be passed on to Python and CSharp wrapper libraries

The issue:

This seems to cause an extra underscrore in the output name which breaks
the imports and .i file. The fix is to set the name conditionally

Notes:

This should not affect users building python modules from the default
python/ directory. Its only the special python3/ extra directory.