Remove deprecated use of distutils, fix for Python 3.12+
This switches to using sysconfig from distutils, which is necessary for
Python 3.12+ since distutils is deprecated and has been removed.

It is necessary to specify the install scheme when a prefix other than
the Python default is used so that changes to the default scheme made by
distributions (e.g. Debian, Fedora) do not produce an incorrect Python
installation directory. For example, Debian patches the default scheme
to prepend the path with '/local', but if a user specifies a prefix of
'/usr/local', then the path using the default scheme would be
'/usr/local/local/...' with a duplicated 'local' directory. Specifying
an unmodified install scheme fixes that.

Signed-off-by: Ryan Volz <ryan.volz@gmail.com>